requirements.txt
flask>=3.0
gunicorn>=21.2
anthropic>=0.40
youtube-transcript-api>=1.0
yt-dlp>=2024.0
apscheduler>=3.10,<4.0
sqlalchemy>=2.0
flask-login>=0.6
flask-limiter>=3.5
python-dotenv>=1.0
werkzeug>=3.0
